Move idempotency to baseclass. Allow facilities implement own store
