+++ app/idempotency.py
@@ -1,4 +1,4 @@
-"""Idempotency store for mutating endpoints (submit_job, update_job).
+"""Idempotency store for mutating API endpoints (submit_job, update_job).
 
 Behaviour by Idempotency-Key header:
   - No header         -> pass through, normal execution every time.
@@ -8,29 +8,23 @@ Behaviour by Idempotency-Key header:
   - Same key, different body -> 422 Unprocessable Entity.
   - Adapter raises    -> lock is released; client may retry safely.
 
-Backing stores:
+The backing store is defined via IRI_IDEMPOTENCY_STORE (see create_store).
+Reference implementations:
   - InMemoryIdempotencyStore  : in-process dict; dev/single-instance only.
-  - RedisIdempotencyStore     : Redis-backed; required for multi-replica.
-
-Select via REDIS_URL env var (see config.py).
+  - RedisIdempotencyStore     : Redis-backed;
 """
 
 import hashlib
+import importlib
 import json
 import logging
 import os
-import time
+from abc import ABC, abstractmethod
 
-import redis.asyncio as aioredis
 from fastapi import HTTPException
 from fastapi.responses import JSONResponse
-from redis.exceptions import WatchError
 
 log = logging.getLogger(__name__)
-
-_LOCK_PREFIX = "LOCKED:"
-_DONE_PREFIX = "DONE:"
-_LOCK_TTL_SECONDS = int(os.environ.get("LOCK_TTL_SECONDS", 60))
 
 
 def build_cache_key(user_id: str, idempotency_key: str, endpoint: str) -> str:
@@ -45,7 +39,39 @@ def build_body_hash(body: dict | None) -> str:
     return hashlib.sha256(raw.encode()).hexdigest()
 
 
-async def run_with_idempotency(store, cache_key: str, body_hash: str, adapter_fn) -> JSONResponse:
+class IdempotencyStore(ABC):
+    """Abstract class for idempotency backing stores.
+    """
+
+    @abstractmethod
+    async def check_and_lock(self, cache_key: str, body_hash: str) -> tuple[str, dict | None, int | None]:
+        """Check state and acquire lock for a new request.
+
+        Returns one of:
+          ("proceed", None, None)              -- new request; caller must call store_result or release_lock
+          ("hit", body_dict, status_int)       -- cached result; return it directly
+          ("conflict", None, None)             -- in-flight; caller should return 409
+          ("fingerprint_mismatch", None, None) -- same key, different body; caller should return 422
+        """
+
+    @abstractmethod
+    async def store_result(self, cache_key: str, body_hash: str, response_body: dict, response_status: int) -> None:
+        """Persist the final response and transition the key to DONE state.
+        Must be owner-checked: no-op if this caller no longer holds the lock.
+        """
+
+    @abstractmethod
+    async def release_lock(self, cache_key: str) -> None:
+        """Release the lock when the adapter raises an exception.
+        Must be owner-checked: no-op if the key no longer holds a LOCKED: value.
+        """
+
+    @abstractmethod
+    async def close(self) -> None:
+        """Release connections and resources at application shutdown."""
+
+
+async def run_with_idempotency(store: IdempotencyStore, cache_key: str, body_hash: str, adapter_fn) -> JSONResponse:
     """Run adapter_fn under idempotency control."""
     action, cached_body, cached_status = await store.check_and_lock(cache_key, body_hash)
 
@@ -66,167 +92,19 @@ async def run_with_idempotency(store, cache_key: str, body_hash: str, adapter_fn
         raise
 
 
-class InMemoryIdempotencyStore:
-    """Simple in-process dict store with monotonic-clock TTL. Only for DEV"""
+def create_store() -> IdempotencyStore:
+    """Return the idempotency store configured via IRI_IDEMPOTENCY_STORE.
 
-    def __init__(self, ttl: int):
-        self._ttl = ttl
-        self._data: dict[str, tuple[str, float]] = {}
-
-    def _get(self, key: str) -> str | None:
-        entry = self._data.get(key)
-        if entry is None:
-            return None
-        value, expires_at = entry
-        if time.monotonic() > expires_at:
-            del self._data[key]
-            return None
-        return value
-
-    def _set(self, key: str, value: str, ttl: int) -> None:
-        self._data[key] = (value, time.monotonic() + ttl)
-
-    def _delete(self, key: str) -> None:
-        self._data.pop(key, None)
-
-    async def check_and_lock(self, cache_key: str, body_hash: str) -> tuple[str, dict | None, int | None]:
-        """Check state and acquire lock for a new request.
-        Possible returns:
-          ("proceed", None, None)              -- new request; caller must call store_result or release_lock
-          ("hit", body_dict, status_int)       -- cached result; return it directly
-          ("conflict", None, None)             -- in-flight; caller should return 409
-          ("fingerprint_mismatch", None, None) -- same key, different body; caller should return 422
-        """
-        value = self._get(cache_key)
-
-        if value is None:
-            self._set(cache_key, f"{_LOCK_PREFIX}{body_hash}", _LOCK_TTL_SECONDS)
-            return ("proceed", None, None)
-
-        if value.startswith(_LOCK_PREFIX):
-            return ("conflict", None, None)
-
-        if value.startswith(_DONE_PREFIX):
-            data = json.loads(value[len(_DONE_PREFIX):])
-            if data["body_hash"] != body_hash:
-                return ("fingerprint_mismatch", None, None)
-            return ("hit", data["response_body"], data["response_status"])
-
-        return ("conflict", None, None)
-
-    async def store_result(self, cache_key: str, body_hash: str, response_body: dict, response_status: int) -> None:
-        """Overwrite the lock entry with the final response and extend TTL.
-
-        Single-threaded asyncio: no await between the check and set, so this is atomic.
-        """
-        value = self._get(cache_key)
-        if value != f"{_LOCK_PREFIX}{body_hash}":
-            return
-        data = {"body_hash": body_hash, "response_body": response_body, "response_status": response_status}
-        self._set(cache_key, f"{_DONE_PREFIX}{json.dumps(data)}", self._ttl)
-
-    async def release_lock(self, cache_key: str) -> None:
-        """Delete the lock.
-
-        Single-threaded asyncio: no await between the check and delete, so this is atomic.
-        """
-        value = self._get(cache_key)
-        if value and value.startswith(_LOCK_PREFIX):
-            self._delete(cache_key)
-
-    async def close(self) -> None:
-        pass
-
-
-class RedisIdempotencyStore:
-    """Redis-backed Idempotency Store"""
-
-    def __init__(self, redis_url: str, ttl: int):
-        self._client = aioredis.from_url(redis_url, decode_responses=True)
-        self._ttl = ttl
-
-    def _rkey(self, cache_key: str) -> str:
-        return f"iri:idem:{cache_key}"
-
-    async def check_and_lock(self, cache_key: str, body_hash: str) -> tuple[str, dict | None, int | None]:
-        rkey = self._rkey(cache_key)
-        lock_value = f"{_LOCK_PREFIX}{body_hash}"
-
-        is_new = await self._client.set(rkey, lock_value, nx=True, ex=_LOCK_TTL_SECONDS)
-        if is_new:
-            return ("proceed", None, None)
-
-        value = await self._client.get(rkey)
-        if value is None:
-            # Key expired between our SET NX and GET; try once more.
-            is_new2 = await self._client.set(rkey, lock_value, nx=True, ex=_LOCK_TTL_SECONDS)
-            if is_new2:
-                return ("proceed", None, None)
-            return ("conflict", None, None)
-
-        if value.startswith(_LOCK_PREFIX):
-            return ("conflict", None, None)
-
-        if value.startswith(_DONE_PREFIX):
-            data = json.loads(value[len(_DONE_PREFIX):])
-            if data["body_hash"] != body_hash:
-                return ("fingerprint_mismatch", None, None)
-            return ("hit", data["response_body"], data["response_status"])
-
-        return ("conflict", None, None)
-
-    async def store_result(self, cache_key: str, body_hash: str, response_body: dict, response_status: int) -> None:
-        """Write DONE only if we still own the lock, using WATCH/MULTI/EXEC optimistic locking.
-
-        If the lock expired and another request acquired it between check_and_lock and here,
-        the WATCH detects the change and the SET is skipped.
-        """
-        rkey = self._rkey(cache_key)
-        expected_lock = f"{_LOCK_PREFIX}{body_hash}"
-        data = {"body_hash": body_hash, "response_body": response_body, "response_status": response_status}
-        done_value = f"{_DONE_PREFIX}{json.dumps(data)}"
-
-        async with self._client.pipeline() as pipe:
-            try:
-                await pipe.watch(rkey)
-                if await pipe.get(rkey) != expected_lock:
-                    await pipe.reset()
-                    return
-                pipe.multi()
-                pipe.set(rkey, done_value, ex=self._ttl)
-                await pipe.execute()
-            except WatchError:
-                pass  # key changed between watch and execute; another request owns it now
-
-    async def release_lock(self, cache_key: str) -> None:
-        """Delete the lock only if it still holds a LOCKED: value, using WATCH/MULTI/EXEC.
-
-        Prevents deleting a lock that expired and was re-acquired by a different request.
-        """
-        rkey = self._rkey(cache_key)
-
-        async with self._client.pipeline() as pipe:
-            try:
-                await pipe.watch(rkey)
-                current = await pipe.get(rkey)
-                if not (current and current.startswith(_LOCK_PREFIX)):
-                    await pipe.reset()
-                    return
-                pipe.multi()
-                pipe.delete(rkey)
-                await pipe.execute()
-            except WatchError:
-                pass  # key changed between watch and execute; leave it alone
-
-    async def close(self) -> None:
-        await self._client.aclose()
-
-
-def create_store(redis_url: str, ttl: int) -> InMemoryIdempotencyStore | RedisIdempotencyStore:
-    """Return Redis-backed store if REDIS_URL is set, in-memory otherwise."""
-    if redis_url:
-        store = RedisIdempotencyStore(redis_url, ttl)
-        log.info("Idempotency store: Redis at %s (TTL %ds)", redis_url, ttl)
-        return store
-    log.info("Idempotency store: in-memory (single-instance; TTL %ds)", ttl)
-    return InMemoryIdempotencyStore(ttl)
+    The named class is imported and instantiated with no arguments — it reads any
+    connection strings or config it needs from env vars.
+    Defaults to app.demo_adapter.InMemoryIdempotencyStore (single-instance; not for production).
+    """
+    class_path = os.environ.get("IRI_IDEMPOTENCY_STORE", "app.demo_adapter.InMemoryIdempotencyStore")
+    parts = class_path.rsplit(".", 1)
+    module = importlib.import_module(parts[0])
+    StoreClass = getattr(module, parts[1])
+    if not issubclass(StoreClass, IdempotencyStore):
+        raise ValueError(f"{class_path} must subclass IdempotencyStore")
+    store = StoreClass()
+    log.info("Idempotency store: %s", class_path)
+    return store
